[fix]mytab 컴포넌트->페이지로분리
